- Show version chip on details page - version chip opens popper with info (classification, expiration) - version action only via button on details page
